@@ -1,30 +1,32 @@
-// Package logger provides a convience function to constructing a logger for use.
+// Package logger provides a convenience function to constructing a logger
+// for use. This is required not just for applications but for testing.
 package logger
 
 import (
-	"fmt"
-	"os"
-
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 )
 
 // New constructs a Sugared Logger that writes to stdout and
-// provides human readable timestamps.
-func New(service string) *zap.SugaredLogger {
+// provides human-readable timestamps.
+func New(service string, outputPaths ...string) (*zap.SugaredLogger, error) {
 	config := zap.NewProductionConfig()
-	config.OutputPaths = []string{"stdout"}
+
 	config.EncoderConfig.EncodeTime = zapcore.ISO8601TimeEncoder
 	config.DisableStacktrace = true
-	config.InitialFields = map[string]interface{}{
+	config.InitialFields = map[string]any{
 		"service": service,
 	}
 
-	log, err := config.Build()
-	if err != nil {
-		fmt.Println(err)
-		os.Exit(1)
+	config.OutputPaths = []string{"stdout"}
+	if outputPaths != nil {
+		config.OutputPaths = outputPaths
 	}
 
-	return log.Sugar()
+	log, err := config.Build(zap.WithCaller(true))
+	if err != nil {
+		return nil, err
+	}
+
+	return log.Sugar(), nil
 }
