@@ -5,49 +5,54 @@ import (
 	"errors"
 )
 
-// ErrInvalidID occurs when an ID is not in a valid form.
-var ErrInvalidID = errors.New("ID is not in its proper form")
-
-// ErrorResponse is the form used for API responses from failures in the API.
-type ErrorResponse struct {
-	Error  string `json:"error"`
-	Fields string `json:"fields,omitempty"`
-}
-
-// RequestError is used to pass an error during the request through the
-// application with web specific context.
-type RequestError struct {
-	Err    error
-	Status int
-	Fields error
-}
-
-// NewRequestError wraps a provided error with an HTTP status code. This
-// function should be used when handlers encounter expected errors.
-func NewRequestError(err error, status int) error {
-	return &RequestError{err, status, nil}
-}
-
-// Error implements the error interface. It uses the default message of the
-// wrapped error. This is what will be shown in the services' logs.
-func (err *RequestError) Error() string {
-	return err.Err.Error()
-}
-
 // FieldError is used to indicate an error with a specific request field.
 type FieldError struct {
 	Field string `json:"field"`
-	Error string `json:"error"`
+	Err   string `json:"error"`
 }
 
 // FieldErrors represents a collection of field errors.
 type FieldErrors []FieldError
 
-// Error implments the error interface.
+// NewFieldsError creates an fields error.
+func NewFieldsError(field string, err error) error {
+	return FieldErrors{
+		{
+			Field: field,
+			Err:   err.Error(),
+		},
+	}
+}
+
+// Error implements the error interface.
 func (fe FieldErrors) Error() string {
 	d, err := json.Marshal(fe)
 	if err != nil {
 		return err.Error()
 	}
 	return string(d)
+}
+
+// Fields returns the fields that failed validation
+func (fe FieldErrors) Fields() map[string]string {
+	m := make(map[string]string)
+	for _, fld := range fe {
+		m[fld.Field] = fld.Err
+	}
+	return m
+}
+
+// IsFieldErrors checks if an error of type FieldErrors exists.
+func IsFieldErrors(err error) bool {
+	var fe FieldErrors
+	return errors.As(err, &fe)
+}
+
+// GetFieldErrors returns a copy of the FieldErrors pointer.
+func GetFieldErrors(err error) FieldErrors {
+	var fe FieldErrors
+	if !errors.As(err, &fe) {
+		return nil
+	}
+	return fe
 }
