@@ -5,9 +5,10 @@ package docker
 import (
 	"bytes"
 	"encoding/json"
+	"fmt"
 	"net"
 	"os/exec"
-	"testing"
+	"strings"
 )
 
 // Container tracks information about the docker container started for tests.
@@ -17,7 +18,7 @@ type Container struct {
 }
 
 // StartContainer starts the specified container for running tests.
-func StartContainer(t *testing.T, image string, port string, args ...string) *Container {
+func StartContainer(image string, port string, args ...string) (*Container, error) {
 	arg := []string{"run", "-P", "-d"}
 	arg = append(arg, args...)
 	arg = append(arg, image)
@@ -26,89 +27,74 @@ func StartContainer(t *testing.T, image string, port string, args ...string) *Co
 	var out bytes.Buffer
 	cmd.Stdout = &out
 	if err := cmd.Run(); err != nil {
-		t.Fatalf("could not start container %s: %v", image, err)
+		return nil, fmt.Errorf("could not start container %s: %w", image, err)
 	}
 
 	id := out.String()[:12]
-
-	cmd = exec.Command("docker", "inspect", id)
-	out.Reset()
-	cmd.Stdout = &out
-	if err := cmd.Run(); err != nil {
-		t.Fatalf("could not inspect container %s: %v", id, err)
+	hostIP, hostPort, err := extractIPPort(id, port)
+	if err != nil {
+		StopContainer(id)
+		return nil, fmt.Errorf("could not extract ip/port: %w", err)
 	}
-
-	var doc []map[string]interface{}
-	if err := json.Unmarshal(out.Bytes(), &doc); err != nil {
-		t.Fatalf("could not decode json: %v", err)
-	}
-
-	ip, randPort := extractIPPort(t, doc, port)
 
 	c := Container{
 		ID:   id,
-		Host: net.JoinHostPort(ip, randPort),
+		Host: net.JoinHostPort(hostIP, hostPort),
 	}
 
-	t.Logf("Image:       %s", image)
-	t.Logf("ContainerID: %s", c.ID)
-	t.Logf("Host:        %s", c.Host)
-
-	return &c
+	return &c, nil
 }
 
 // StopContainer stops and removes the specified container.
-func StopContainer(t *testing.T, id string) {
+func StopContainer(id string) error {
 	if err := exec.Command("docker", "stop", id).Run(); err != nil {
-		t.Fatalf("could not stop container: %v", err)
+		return fmt.Errorf("could not stop container: %w", err)
 	}
-	t.Log("Stopped:", id)
 
 	if err := exec.Command("docker", "rm", id, "-v").Run(); err != nil {
-		t.Fatalf("could not remove container: %v", err)
+		return fmt.Errorf("could not remove container: %w", err)
 	}
-	t.Log("Removed:", id)
+
+	return nil
 }
 
 // DumpContainerLogs outputs logs from the running docker container.
-func DumpContainerLogs(t *testing.T, id string) {
+func DumpContainerLogs(id string) []byte {
 	out, err := exec.Command("docker", "logs", id).CombinedOutput()
 	if err != nil {
-		t.Fatalf("could not log container: %v", err)
+		return nil
 	}
-	t.Logf("Logs for %s\n%s:", id, out)
+	return out
 }
 
-func extractIPPort(t *testing.T, doc []map[string]interface{}, port string) (string, string) {
-	nw, exists := doc[0]["NetworkSettings"]
-	if !exists {
-		t.Fatal("could not get network settings")
-	}
-	ports, exists := nw.(map[string]interface{})["Ports"]
-	if !exists {
-		t.Fatal("could not get network ports settings")
-	}
-	tcp, exists := ports.(map[string]interface{})[port+"/tcp"]
-	if !exists {
-		t.Fatal("could not get network ports/tcp settings")
-	}
-	list, exists := tcp.([]interface{})
-	if !exists {
-		t.Fatal("could not get network ports/tcp list settings")
+func extractIPPort(id string, port string) (hostIP string, hostPort string, err error) {
+	tmpl := fmt.Sprintf("[{{range $k,$v := (index .NetworkSettings.Ports \"%s/tcp\")}}{{json $v}}{{end}}]", port)
+
+	cmd := exec.Command("docker", "inspect", "-f", tmpl, id)
+	var out bytes.Buffer
+	cmd.Stdout = &out
+	if err := cmd.Run(); err != nil {
+		return "", "", fmt.Errorf("could not inspect container %s: %w", id, err)
 	}
 
-	var hostIP string
-	var hostPort string
-	for _, l := range list {
-		data, exists := l.(map[string]interface{})
-		if !exists {
-			t.Fatal("could not get network ports/tcp list data")
-		}
-		hostIP = data["HostIp"].(string)
-		if hostIP != "::" {
-			hostPort = data["HostPort"].(string)
+	// When IPv6 is turned on with Docker.
+	// Got  [{"HostIp":"0.0.0.0","HostPort":"49190"}{"HostIp":"::","HostPort":"49190"}]
+	// Need [{"HostIp":"0.0.0.0","HostPort":"49190"},{"HostIp":"::","HostPort":"49190"}]
+	data := strings.ReplaceAll(out.String(), "}{", "},{")
+
+	var docs []struct {
+		HostIP   string
+		HostPort string
+	}
+	if err := json.Unmarshal([]byte(data), &docs); err != nil {
+		return "", "", fmt.Errorf("could not decode json: %w", err)
+	}
+
+	for _, doc := range docs {
+		if doc.HostIP != "::" {
+			return doc.HostIP, doc.HostPort, nil
 		}
 	}
 
-	return hostIP, hostPort
+	return "", "", fmt.Errorf("could not locate ip/port")
 }
