@@ -1,30 +1,28 @@
 package web
 
 import (
-	"github.com/pkg/errors"
+	"errors"
 )
 
-// shutdown is a type used to help with the graceful termination of the service.
-type shutdown struct {
+// shutdownError is a type used to help with the graceful termination of the service.
+type shutdownError struct {
 	Message string
 }
 
 // NewShutdownError returns an error that causes the framework to signal
 // a graceful shutdown.
 func NewShutdownError(message string) error {
-	return &shutdown{message}
+	return &shutdownError{message}
 }
 
 // Error is the implementation of the error interface.
-func (s *shutdown) Error() string {
-	return s.Message
+func (se *shutdownError) Error() string {
+	return se.Message
 }
 
 // IsShutdown checks to see if the shutdown error is contained
 // in the specified error value.
 func IsShutdown(err error) bool {
-	if _, ok := errors.Cause(err).(*shutdown); ok {
-		return true
-	}
-	return false
+	var se *shutdownError
+	return errors.As(err, &se)
 }
