@@ -3,6 +3,7 @@ package web
 
 import (
 	"context"
+	"errors"
 	"net/http"
 	"os"
 	"syscall"
@@ -13,20 +14,7 @@ import (
 	"go.opentelemetry.io/otel/trace"
 )
 
-// ctxKey represents the type of value for the context key.
-type ctxKey int
-
-// KeyValues is how request values are stored/retrieved.
-const KeyValues ctxKey = 1
-
-// Values represent state for each request.
-type Values struct {
-	TraceID    string
-	Now        time.Time
-	StatusCode int
-}
-
-// A Handler is a type that handles an http request within our own little mini
+// A Handler is a type that handles a http request within our own little mini
 // framework.
 type Handler func(ctx context.Context, w http.ResponseWriter, r *http.Request) error
 
@@ -38,16 +26,17 @@ type App struct {
 	otmux    http.Handler
 	shutdown chan os.Signal
 	mw       []Middleware
+	tracer   trace.Tracer
 }
 
 // NewApp creates an App value that handle a set of routes for the application.
-func NewApp(shutdown chan os.Signal, mw ...Middleware) *App {
+func NewApp(shutdown chan os.Signal, tracer trace.Tracer, mw ...Middleware) *App {
 
 	// Create an OpenTelemetry HTTP Handler which wraps our router. This will start
 	// the initial span and annotate it with information about the request/response.
 	//
 	// This is configured to use the W3C TraceContext standard to set the remote
-	// parent if an client request includes the appropriate headers.
+	// parent if a client request includes the appropriate headers.
 	// https://w3c.github.io/trace-context/
 
 	mux := httptreemux.NewContextMux()
@@ -57,10 +46,11 @@ func NewApp(shutdown chan os.Signal, mw ...Middleware) *App {
 		otmux:    otelhttp.NewHandler(mux, "request"),
 		shutdown: shutdown,
 		mw:       mw,
+		tracer:   tracer,
 	}
 }
 
-// SignalShutdown is used to gracefully shutdown the app when an integrity
+// SignalShutdown is used to gracefully shut down the app when an integrity
 // issue is identified.
 func (a *App) SignalShutdown() {
 	a.shutdown <- syscall.SIGTERM
@@ -69,45 +59,80 @@ func (a *App) SignalShutdown() {
 // ServeHTTP implements the http.Handler interface. It's the entry point for
 // all http traffic and allows the opentelemetry mux to run first to handle
 // tracing. The opentelemetry mux then calls the application mux to handle
-// application traffic. This was setup on line 58 in the NewApp function.
+// application traffic. This was set up on line 44 in the NewApp function.
 func (a *App) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	a.otmux.ServeHTTP(w, r)
 }
 
 // Handle sets a handler function for a given HTTP method and path pair
 // to the application server mux.
-func (a *App) Handle(method string, path string, handler Handler, mw ...Middleware) {
-
-	// First wrap handler specific middleware around this handler.
+func (a *App) Handle(method string, group string, path string, handler Handler, mw ...Middleware) {
 	handler = wrapMiddleware(mw, handler)
-
-	// Add the application's general middleware to the handler chain.
 	handler = wrapMiddleware(a.mw, handler)
 
-	// The function to execute for each request.
 	h := func(w http.ResponseWriter, r *http.Request) {
-
-		// Pull the context from the request and
-		// use it as a separate parameter.
 		ctx := r.Context()
 
-		// Capture the parent request span from the context.
 		span := trace.SpanFromContext(ctx)
 
-		// Set the context with the required values to
-		// process the request.
 		v := Values{
 			TraceID: span.SpanContext().TraceID().String(),
-			Now:     time.Now(),
+			Tracer:  a.tracer,
+			Now:     time.Now().UTC(),
 		}
-		ctx = context.WithValue(ctx, KeyValues, &v)
+		ctx = context.WithValue(ctx, key, &v)
 
-		// Call the wrapped handler functions.
 		if err := handler(ctx, w, r); err != nil {
-			a.SignalShutdown()
-			return
+			if validateShutdown(err) {
+				a.SignalShutdown()
+				return
+			}
 		}
 	}
 
-	a.mux.Handle(method, path, h)
+	finalPath := path
+	if group != "" {
+		finalPath = "/" + group + path
+	}
+	a.mux.Handle(method, finalPath, h)
+}
+
+// validateShutdown validates the error for special conditions that do not
+// warrant an actual shutdown by the system.
+func validateShutdown(err error) bool {
+
+	// Ignore syscall.EPIPE and syscall.ECONNRESET errors which occurs
+	// when a write operation happens on the http.ResponseWriter that
+	// has simultaneously been disconnected by the client (TCP
+	// connections is broken). For instance, when large amounts of
+	// data is being written or streamed to the client.
+	// https://blog.cloudflare.com/the-complete-guide-to-golang-net-http-timeouts/
+	// https://gosamples.dev/broken-pipe/
+	// https://gosamples.dev/connection-reset-by-peer/
+
+	switch {
+	case errors.Is(err, syscall.EPIPE):
+
+		// Usually, you get the broken pipe error when you write to the connection after the
+		// RST (TCP RST Flag) is sent.
+		// The broken pipe is a TCP/IP error occurring when you write to a stream where the
+		// other end (the peer) has closed the underlying connection. The first write to the
+		// closed connection causes the peer to reply with an RST packet indicating that the
+		// connection should be terminated immediately. The second write to the socket that
+		// has already received the RST causes the broken pipe error.
+		return false
+
+	case errors.Is(err, syscall.ECONNRESET):
+
+		// Usually, you get connection reset by peer error when you read from the
+		// connection after the RST (TCP RST Flag) is sent.
+		// The connection reset by peer is a TCP/IP error that occurs when the other end (peer)
+		// has unexpectedly closed the connection. It happens when you send a packet from your
+		// end, but the other end crashes and forcibly closes the connection with the RST
+		// packet instead of the TCP FIN, which is used to close a connection under normal
+		// circumstances.
+		return false
+	}
+
+	return true
 }
